+++ scripts/changelog/eslint.config.mjs
@@ -1,0 +1,22 @@
+import eslint from '@eslint/js';
+import prettier from 'eslint-config-prettier';
+import { defineConfig, globalIgnores } from 'eslint/config';
+import tseslint from 'typescript-eslint';
+
+export default defineConfig([
+    globalIgnores(['coverage/', 'reports/']),
+    {
+        files: ['**/*.ts'],
+        extends: [
+            eslint.configs.recommended,
+            tseslint.configs.recommendedTypeChecked,
+            tseslint.configs.stylisticTypeChecked,
+        ],
+        languageOptions: {
+            parserOptions: {
+                project: ['./tsconfig.json', './tsconfig.spec.json'],
+            },
+        },
+    },
+    prettier,
+]);